factor out schema into separate subproject
